fix: refactor route 53 and add extra test
add an extra test to check hosted zone and recordset creation
on same transaction
