+++ src/modules/0.0.23/aws_route53/mappers/index.ts
@@ -1,0 +1,2 @@
+export * from './hosted_zone';
+export * from './resource_record_set';
